gestion_personas/urls.py
from django.urls import path
from . import views

urlpatterns = [
    path('', views.lista_personas, name='lista_personas'),
    path('consulta/', views.consulta_sql, name='consulta_sql'),
    path('agregar/', views.agregar_persona, name='agregar_persona'),
    path('editar/<int:id>/', views.editar_persona, name='editar_persona'),
    path('eliminar/<int:id>/', views.eliminar_persona, name='eliminar_persona'),
]
